提交

+++ ZHLZ/CoreBusiness/Home/View/Cell/HomeInfoStatistics/ZHLZHomeInfoStatisticsTableViewCell.m
@@ -33,10 +33,10 @@
     _homeInfoStatisticsModel = homeInfoStatisticsModel;
     
     self.brigadeLabel.text = _homeInfoStatisticsModel.orgname?:@"";
-    self.queryKilometersCountLabel.text = [NSString stringWithFormat:@"%ld", _homeInfoStatisticsModel.mapMileage];
-    self.queryPeopleCountLabel.text = [NSString stringWithFormat:@"%ld", _homeInfoStatisticsModel.foundnum];
-    self.findProblemCountLabel.text = [NSString stringWithFormat:@"%ld", _homeInfoStatisticsModel.findnum];
-    self.closedProblemCountLabel.text = [NSString stringWithFormat:@"%ld", _homeInfoStatisticsModel.closenum];
+    self.queryKilometersCountLabel.text = [NSString stringWithFormat:@"%ld", (long)_homeInfoStatisticsModel.mapMileage];
+    self.queryPeopleCountLabel.text = [NSString stringWithFormat:@"%ld", (long)_homeInfoStatisticsModel.foundnum];
+    self.findProblemCountLabel.text = [NSString stringWithFormat:@"%ld", (long)_homeInfoStatisticsModel.findnum];
+    self.closedProblemCountLabel.text = [NSString stringWithFormat:@"%ld", (long)_homeInfoStatisticsModel.closenum];
 }
 
 @end
